// packages/source-sync/src/lambda-modules/sync-boundary/infrastructure/adapter/out/db/typeorm/fields/mt-grid-event-columns.enum.ts
export enum MtGridEventColumns {
  ID = 'ID',
  DOMAIN_ID = 'DOMAIN_ID',
  DOCUMENT_ID = 'DOCUMENT_ID',
  NEXUS_ANOMALY_NUMBER = 'NEXUS_ANOMALY_NUMBER',
  SAP_NOTIFICATION_NUMBER = 'SAP_NOTIFICATION_NUMBER',
  SAP_ORDER_NUMBER = 'SAP_ORDER_NUMBER',
  EVENT_TYPE = 'EVENT_TYPE',
  DETECTION_DATE = 'DETECTION_DATE',
  DETECTION_COMPANY = 'DETECTION_COMPANY',
  DETECTION_TECHNICIAN = 'DETECTION_TECHNICIAN',
  PRIORITY = 'PRIORITY',
  TECHNICAL_LOCATION = 'TECHNICAL_LOCATION',
  EVENT_STATUS = 'EVENT_STATUS',
  REPAIR_STATUS = 'REPAIR_STATUS',
  REPAIR_STATUS_UPDATED_AT = 'REPAIR_STATUS_UPDATED_AT',
  POWER_BI_STATUS = 'POWER_BI_STATUS',
  HISTORICAL_AT = 'HISTORICAL_AT',
  AFFECTS_SUPPLY = 'AFFECTS_SUPPLY',
  INTERVENING_TEAMS = 'INTERVENING_TEAMS',
  REGIONAL_DESK_REMARKS = 'REGIONAL_DESK_REMARKS',
  EXECUTION_REMARKS = 'EXECUTION_REMARKS',
  MUNICIPAL_PERMIT = 'MUNICIPAL_PERMIT',
  DAMAGE_BY_THIRD_PARTY = 'DAMAGE_BY_THIRD_PARTY',
  FAILURES_COUNT = 'FAILURES_COUNT',
  FAILURE_ADDRESS = 'FAILURE_ADDRESS',
  GEO_LOCATION = 'GEO_LOCATION',
  MATERIAL_TRANSFORMER_THERMO = 'MATERIAL_TRANSFORMER_THERMO',
  MATERIAL_TRANSFORMER_COLD = 'MATERIAL_TRANSFORMER_COLD',
  MATERIAL_UNIPOLAR_CABLE = 'MATERIAL_UNIPOLAR_CABLE',
  MATERIAL_TERMINAL = 'MATERIAL_TERMINAL',
  MATERIAL_TERNA_METERS = 'MATERIAL_TERNA_METERS',
  CRITICALITY = 'CRITICALITY',
  INSTALLATION = 'INSTALLATION',
  FAILURE_TYPE = 'FAILURE_TYPE',
  FEEDER = 'FEEDER',
  SUBSTATION_CODE = 'SUBSTATION_CODE',
  SUBSTATION_NAME = 'SUBSTATION_NAME',
  DELIVERY_RESPONSIBLE_ROLE = 'DELIVERY_RESPONSIBLE_ROLE',
  DELIVERY_RESPONSIBLE_OTHER_TEXT = 'DELIVERY_RESPONSIBLE_OTHER_TEXT',
  NORMALIZATION_TASK_REFERENCE = 'NORMALIZATION_TASK_REFERENCE',
  PROTOCOL_NUMBER = 'PROTOCOL_NUMBER',
  PROTOCOL_STARTED_AT = 'PROTOCOL_STARTED_AT',
  PROTOCOL_ENDED_AT = 'PROTOCOL_ENDED_AT',
  REPAIR_STARTED_AT = 'REPAIR_STARTED_AT',
  ESTIMATED_FINISHED_AT = 'ESTIMATED_FINISHED_AT',
  ACTUAL_FINISHED_AT = 'ACTUAL_FINISHED_AT',
  MANUAL_OVERRIDES = 'MANUAL_OVERRIDES',
  DELETED_AT = 'DELETED_AT',
  VERSION = 'VERSION',
  CREATED_AT = 'CREATED_AT',
  UPDATED_AT = 'UPDATED_AT',
}